pyproject.toml
[project]
name = "ragformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=24.10.0",
    "colorama>=0.4.6",
    "diagrams>=0.23.4",
    "llama-index-embeddings-azure-openai>=0.2.5",
    "llama-index-llms-anthropic>=0.3.6",
    "llama-index-llms-azure-openai>=0.2.2",
    "llama-index-llms-groq>=0.2.0",
    "llama-index-llms-ollama>=0.3.4",
    "llama-index-llms-together>=0.2.0",
    "llama-index-utils-workflow>=0.2.1",
    "llama-index>=0.11.17",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "together>=1.3.1",
    "toolhouse>=1.3.0",
]
